requirements.txt
torch
pillow
numpy
diffusers
transformers
accelerate
safetensors
fastapi
uvicorn[standard]
python-multipart